GlitchGarden : - Menu button - Gravestone animation fix - Game balance tunning - Difficulty option integration

+++ Unity4/GlitchGarder/Assets/Scripts/OptionsController.cs
@@ -26,7 +26,7 @@ public class OptionsController : MonoBehaviour {
 	
 	public void SaveAndExit() {
 		PlayerPrefsManager.SetMasterVolume(volumeSlider.value);
-		PlayerPrefsManager.SetDifficulty(difficultySlider.value);
+		PlayerPrefsManager.SetDifficulty((int)difficultySlider.value);
 		
 		levelManager.LoadLevel("01a MainMenu");
 	}
